// src/types/pokemon.ts
export type PokemonType = {
  id: string
  name: string;
  number: number;
  photo: string;
  type: string;
  description: string;
  height: string;
  weight: string;
  genderRatio1: string;
  genderRatio2: string;
  abilities: string;
  eggGroups: string;
  evolutionDesc: string;
  evolutionPhoto: string;
};